move group subpages to pages folder
